refactor(ui): reorganize src/ into components/, context/, and utils/
