+++ ui/src/components/panels/SettingsPanel.tsx
@@ -1,4 +1,4 @@
-import { useSettings } from "./SettingsContext";
+import { useSettings } from "../../context/SettingsContext";
 
 const THINKING_LEVELS = ["off", "minimal", "low", "medium", "high", "xhigh"] as const;
