+++ ui/src/components/sidebar/FileTree.tsx
