+++ ui/src/utils/context-panel-utils.ts
@@ -1,4 +1,4 @@
-import type { ContextGroup, ContextFile } from "./types";
+import type { ContextGroup, ContextFile } from "../types";
 
 export function countContextFiles(groups: ContextGroup[]): number {
   return groups.reduce((n, g) => n + g.files.length, 0);
